requirements.txt
deepgram-sdk==3.*
python-dotenv
jiwer
tf-keras
torch
transformers